fixup! KAFKA-14226: feat(connect:transform): Introduce FieldPath abstraction
Moving tests as suggested [1]

[1]: https://github.com/apache/kafka/pull/15379#discussion_r1571018390
